added mostly comments and I suggest how to loop through the buffers without the actual length

+++ kernel/src/hil/radio_nrf51dk.rs
@@ -3,10 +3,9 @@ use returncode::ReturnCode;
 // Defines the Interface between Capsules and Chips
 pub trait RadioDriver {
     fn init(&self);
-
     fn send(&self);
     fn receive(&self);
-    fn transmit(&self, dest: u16, tx_data: &'static mut [u8], tx_len: u8) -> ReturnCode;
+    fn transmit(&self, dest: u16, tx_data: &'static mut [u8], tx_len: usize) -> ReturnCode;
     // ADD MORE LATER
     fn flash_leds(&self);
     fn set_channel(&self, ch: usize);
